fix: apply all PR review feedback — boundary checks, idempotent migration, event type, worker race condition

+++ src/Services/ExpressRecipe.InventoryService/Services/MealDelayStorageSubscriber.cs
@@ -47,7 +47,7 @@ public sealed class MealDelayStorageSubscriber : IHostedService
         }
 
         double daysDelta = (evt.NewPlannedDate.Value - evt.OldPlannedDate.Value).TotalDays;
-        if (daysDelta < 2)
+        if ((int)daysDelta < 2)
         {
             return;
         }
